move admin-status helper out of diff folder and fix localhost handling
